requirements.txt
# --- Core Deep Learning ---
torch
torchvision

# --- Image/Video Processing ---
opencv-python-headless
numpy

# --- Utilities ---
gdown
PyYAML
tqdm
scikit-image

# --- Flask for potential server deployment ---
Flask